SocketPairTest::PassMessage works but not PassBinary
SocketPairTest::PassBinary is not working. It seems to be because the default istream tries to extract ASCII. It also tokenizes the input if there is whitespace. All of those things are undesireable for data stream processing. One solution is to override fstream types, but this means that the filebuf type must be used for rdbuf(), but filebuf does not have a setter for rdbuf(), only a getter, which prevents copy assignment.