Add integer file descriptor into fstream. Use Map for mode-string conversion.
If the filebuf was created using an integer file descriptor, a corresponding FILE * is created. Otherwise, if the FILE * is opened, then the corresponding integer file descriptor is cached using the fileno() call.

Also, previously a linear search was used to convert std::ios_base::openmode to string. I might switch the string part to const char * in the next commit.